chore(server): move router to routes/
- routes/api can be used to split api routes up
- controllers moved to routes/controllers
- router.js -> routes/index.js
